refactor: 패키지, 클래스명에 reservation 붙이도록 통일
