+++ src/main/java/roomescape/exception/reservationtime/DuplicatedReservationTimeException.java
@@ -1,10 +1,10 @@
-package roomescape.exception.time;
+package roomescape.exception.reservationtime;
 
 import org.springframework.http.HttpStatus;
 import roomescape.exception.common.RoomescapeException;
 
-public class DuplicatedTimeException extends RoomescapeException {
-    public DuplicatedTimeException() {
+public class DuplicatedReservationTimeException extends RoomescapeException {
+    public DuplicatedReservationTimeException() {
         super("해당 예약 시간이 이미 존재합니다.", HttpStatus.CONFLICT);
     }
 }
