+++ src/main/java/roomescape/exception/reservationtime/NotFoundReservationTimeException.java
@@ -1,10 +1,10 @@
-package roomescape.exception.time;
+package roomescape.exception.reservationtime;
 
 import org.springframework.http.HttpStatus;
 import roomescape.exception.common.RoomescapeException;
 
-public class NotFoundTimeException extends RoomescapeException {
-    public NotFoundTimeException() {
+public class NotFoundReservationTimeException extends RoomescapeException {
+    public NotFoundReservationTimeException() {
         super("존재하지 않는 시간입니다.", HttpStatus.NOT_FOUND);
     }
 }
