+++ src/main/java/roomescape/exception/reservationwaiting/InvalidDateTimeReservationWaitingException.java
@@ -3,8 +3,8 @@ package roomescape.exception.reservationwaiting;
 import org.springframework.http.HttpStatus;
 import roomescape.exception.common.RoomescapeException;
 
-public class InvalidDateTimeWaitingException extends RoomescapeException {
-    public InvalidDateTimeWaitingException() {
+public class InvalidDateTimeReservationWaitingException extends RoomescapeException {
+    public InvalidDateTimeReservationWaitingException() {
         super("지나간 날짜와 시간에 대한 예약 대기입니다.", HttpStatus.BAD_REQUEST);
     }
 }
